style: rename file with tree manipulation logic
