+++ src/utils/zon/index.ts
@@ -2,4 +2,4 @@ export * from "./types.ts";
 export * from "./node.ts";
 export * from "./path.ts";
 export * from "./stats.ts";
-export * from "./zon.ts";
+export * from "./tree.ts";
